+++ tests/adapter/fixtures/TestClassNoDoc.js
@@ -1,0 +1,10 @@
+class TestClassNoDoc {
+  constructor (value = 0) {
+    this._value = value
+  }
+  getValue () {
+    return this._value
+  }
+}
+
+module.exports = TestClassNoDoc